Restructure pointcloud folder
