+++ viewer/src/__tests__/viewer/public/RevealManagerBase.test.ts
@@ -7,7 +7,7 @@ import * as THREE from 'three';
 import { RevealManagerBase } from '@/public/RevealManagerBase';
 import { MaterialManager } from '@/datamodels/cad/MaterialManager';
 import { CadManager } from '@/datamodels/cad/CadManager';
-import { PointCloudManager } from '@/datamodels/pointcloud/internal/PointCloudManager';
+import { PointCloudManager } from '@/datamodels/pointcloud/PointCloudManager';
 
 describe('RevealManagerBase', () => {
   const mockCadManager: Omit<CadManager<number>, ''> = {
