+++ viewer/src/datamodels/pointcloud/PointCloudMetadata.ts
