+++ viewer/src/datamodels/pointcloud/PotreeGroupWrapper.ts
