+++ viewer/src/public/RevealManager.ts
@@ -18,9 +18,9 @@ import { CachedRepository } from '@/datamodels/cad/sector/CachedRepository';
 import { CadModelUpdateHandler } from '@/datamodels/cad/CadModelUpdateHandler';
 import { CadManager } from '@/datamodels/cad/CadManager';
 import { ModelNodeAppearance, CadNode } from '@/datamodels/cad';
-import { PointCloudMetadataRepository } from '@/datamodels/pointcloud/internal/PointCloudMetadataRepository';
-import { PointCloudFactory } from '@/datamodels/pointcloud/internal/PointCloudFactory';
-import { PointCloudManager } from '@/datamodels/pointcloud/internal/PointCloudManager';
+import { PointCloudMetadataRepository } from '@/datamodels/pointcloud/PointCloudMetadataRepository';
+import { PointCloudFactory } from '@/datamodels/pointcloud/PointCloudFactory';
+import { PointCloudManager } from '@/datamodels/pointcloud/PointCloudManager';
 
 type CdfModelIdentifier = { modelRevision: IdEither; format: File3dFormat };
 export class RevealManager extends RevealManagerBase<CdfModelIdentifier> {
