+++ viewer/src/public/migration/Cognite3DViewer.ts
@@ -34,9 +34,9 @@ import { CadMetadataParser } from '@/datamodels/cad/parsers/CadMetadataParser';
 import { CadModelFactory } from '@/datamodels/cad/CadModelFactory';
 import { ByVisibilityGpuSectorCuller } from '@/internal';
 import { CadModelUpdateHandler } from '@/datamodels/cad/CadModelUpdateHandler';
-import { PointCloudManager } from '@/datamodels/pointcloud/internal/PointCloudManager';
-import { PointCloudMetadataRepository } from '@/datamodels/pointcloud/internal/PointCloudMetadataRepository';
-import { PointCloudFactory } from '@/datamodels/pointcloud/internal/PointCloudFactory';
+import { PointCloudManager } from '@/datamodels/pointcloud/PointCloudManager';
+import { PointCloudMetadataRepository } from '@/datamodels/pointcloud/PointCloudMetadataRepository';
+import { PointCloudFactory } from '@/datamodels/pointcloud/PointCloudFactory';
 
 export interface RelativeMouseEvent {
   offsetX: number;
